% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_convert_font_size.R
\name{dfr_convert_font_size}
\alias{dfr_convert_font_size}
\title{Convert font size}
\usage{
dfr_convert_font_size(text_size_base = 12, text_size_rel = 0.9166667)
}
\arguments{
\item{text_size_base}{The base text size in points (pt). Defaults to 12 pts.}

\item{text_size_rel}{The default relative size for text elements.}
}
\description{
Convert font size
}
