% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_theme.R
\name{dfr_theme}
\alias{dfr_theme}
\title{Apply custom theme}
\usage{
dfr_theme(
  text_size_base = 12,
  text_color_primary = "#757575",
  axis_line_color = "#E0E0E0",
  line_size_base = text_size_base/22,
  plot_margin = 0.25
)
}
\arguments{
\item{text_size_base}{The base text size in points (pt). Defaults to 12 pts.}

\item{text_color_primary}{The text color. Defaults to #757575, except
for plot title and caption.}

\item{axis_line_color}{The axis line color. Defaults to #E0E0E0.}

\item{line_size_base}{The base size for lines.}

\item{plot_margin}{The base size for plot margins on the top, right, bottom
and left sides in cm. Defaults to 0.25 cm.}
}
\description{
Apply custom theme
}
